requirements.txt
numpy==1.15.1
Pillow==5.3.0
scipy==1.1.0
tensorflow-gpu==1.13.1
wget==3.2
seaborn==0.9.0
easydict==1.9
opencv-python==4.1.0.25
